+++ SensingKitLib/src/main/java/org/sensingkit/sensingkitlib/model/data/SKAccelerometerData.java
@@ -19,13 +19,46 @@
  * along with SensingKit-Android.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.sensingkit.sensingkitlib;
+package org.sensingkit.sensingkitlib.model.data;
 
-import org.sensingkit.sensingkitlib.model.data.SKDataInterface;
-import org.sensingkit.sensingkitlib.modules.SKSensorModuleType;
+import java.util.Locale;
 
-public interface SKSensorDataListener {
+public class SKAccelerometerData extends SKAbstractData {
 
-    void onDataReceived(final SKSensorModuleType moduleType, final SKDataInterface moduleData);
+    @SuppressWarnings("unused")
+    private static final String TAG = "SKAccelerometerData";
+
+    protected final float x;
+    protected final float y;
+    protected final float z;
+
+    public SKAccelerometerData(long timestamp, float x, float y, float z) {
+
+        super(timestamp);
+
+        this.x = x;
+        this.y = y;
+        this.z = z;
+    }
+
+    @Override
+    public String getDataInCSV() {
+        return String.format(Locale.US, "%d,%f,%f,%f", this.timestamp, this.x, this.y, this.z);
+    }
+
+    @SuppressWarnings("unused")
+    public float getX() {
+        return this.x;
+    }
+
+    @SuppressWarnings("unused")
+    public float getY() {
+        return this.y;
+    }
+
+    @SuppressWarnings("unused")
+    public float getZ() {
+        return this.z;
+    }
 
 }