+++ SensingKitLib/src/main/java/org/sensingkit/sensingkitlib/model/data/SKStepCounterData.java
@@ -19,13 +19,32 @@
  * along with SensingKit-Android.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.sensingkit.sensingkitlib;
+package org.sensingkit.sensingkitlib.model.data;
 
-import org.sensingkit.sensingkitlib.model.data.SKDataInterface;
-import org.sensingkit.sensingkitlib.modules.SKSensorModuleType;
+import java.util.Locale;
 
-public interface SKSensorDataListener {
+public class SKStepCounterData extends SKAbstractData {
 
-    void onDataReceived(final SKSensorModuleType moduleType, final SKDataInterface moduleData);
+    @SuppressWarnings("unused")
+    private static final String TAG = "SKStepCounterData";
+
+    protected final float steps;
+
+    public SKStepCounterData(long timestamp, float steps) {
+
+        super(timestamp);
+
+        this.steps = steps;
+    }
+
+    @Override
+    public String getDataInCSV() {
+        return String.format(Locale.US, "%d,%f", this.timestamp, this.steps);
+    }
+
+    @SuppressWarnings("unused")
+    public float getSteps() {
+        return this.steps;
+    }
 
 }