+++ SensingKitLib/src/main/java/org/sensingkit/sensingkitlib/modules/SKStepDetector.java
@@ -19,13 +19,35 @@
  * along with SensingKit-Android.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.sensingkit.sensingkitlib;
+package org.sensingkit.sensingkitlib.modules;
 
-import org.sensingkit.sensingkitlib.model.data.SKDataInterface;
-import org.sensingkit.sensingkitlib.modules.SKSensorModuleType;
+import android.content.Context;
+import android.hardware.SensorEvent;
 
-public interface SKSensorDataListener {
+import org.sensingkit.sensingkitlib.SKException;
+import org.sensingkit.sensingkitlib.model.data.SKAbstractData;
+import org.sensingkit.sensingkitlib.model.data.SKStepDetectorData;
 
-    void onDataReceived(final SKSensorModuleType moduleType, final SKDataInterface moduleData);
+public class SKStepDetector extends SKAbstractNativeSensorModule {
+
+    @SuppressWarnings("unused")
+    private static final String TAG = "SKStepDetector";
+
+    public SKStepDetector(final Context context) throws SKException {
+        super(context, SKSensorModuleType.STEP_DETECTOR);
+    }
+
+    @Override
+    protected SKAbstractData buildData(SensorEvent event)
+    {
+        return new SKStepDetectorData(System.currentTimeMillis());
+    }
+
+    @Override
+    protected boolean shouldPostSensorData(SKAbstractData data) {
+
+        // Always post sensor data
+        return true;
+    }
 
 }