+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-### 0.1.1
-  - Added support for Bluetooth sensor module
+### 0.2.0
+  - Added support for Bluetooth sensor module.
   - DataInterface Method getDataInString() has been renamed to getDataInCSV()
+  - Add SK prefix in all source files
 
 ### 0.1.0
   - Initial Release